+++ go/adbc/driver/flightsql/flightsql_driver.go
@@ -79,7 +79,7 @@ type driverImpl struct {
 // NewDriver creates a new Flight SQL driver using the given Arrow allocator.
 func NewDriver(alloc memory.Allocator) adbc.Driver {
 	info := driverbase.DefaultDriverInfo("Flight SQL")
-	return &driverImpl{DriverImplBase: driverbase.NewDriverImplBase(info, alloc)}
+	return driverbase.NewDriver(&driverImpl{DriverImplBase: driverbase.NewDriverImplBase(info, alloc)})
 }
 
 func (d *driverImpl) NewDatabase(opts map[string]string) (adbc.Database, error) {
@@ -119,5 +119,5 @@ func (d *driverImpl) NewDatabase(opts map[string]string) (adbc.Database, error) 
 		return nil, err
 	}
 
-	return db, nil
+	return driverbase.NewDatabase(db), nil
 }
